+++ DereGuide/Extension/UIColorExtension.swift
@@ -44,6 +44,7 @@ extension UIColor {
     static let normal = UIColor(red: 1.0 * 253 / 255, green: 1.0 * 186 / 255, blue: 1.0 * 80 / 255, alpha: 1)
     static let limited = UIColor(red: 1.0 * 236 / 255, green: 1.0 * 103 / 255, blue: 1.0 * 105 / 255, alpha: 1)
     static let cinfes = UIColor(red: 1.0 * 25 / 255, green: 1.0 * 154 / 255, blue: 1.0 * 218 / 255, alpha: 1)
+    static let premium = UIColor(hexString: "EB2537")
     
     static let flick4 = UIColor(hexString: "3568D5")
     static let flick3 = UIColor(hexString: "7CD8AA")